Don't dynamically send to latest worker
I thought that when the RIO connected the reason it wasn't getting messages was because its identity had changed, but it was actually just because the ROUTER hadn't `recv`'d to signal it was ready. This makes our implementation less "hacky" because we are always sending to the RIO and not just the latest process that pings us.